API/requirements.txt
Django==2.2.7
psycopg2
djangorestframework